major update and bug fixes

+++ include/stack.h
@@ -8,13 +8,13 @@
 typedef struct
 {
     int top;
-    int data[SIZE];
+    double data[SIZE];
 
 } Stack;
 
-void push(Stack *, int);
+void push(Stack *, double);
 
-int pop(Stack *);
+double pop(Stack *);
 
 int isFull(Stack *);
